+++ isatty_windows.go
@@ -10,8 +10,8 @@ import (
 var kernel32 = syscall.NewLazyDLL("kernel32.dll")
 var procGetConsoleMode = kernel32.NewProc("GetConsoleMode")
 
-func IsTerminal(fd int) bool {
+func IsTerminal(fd uintptr) bool {
 	var st uint32
-	r, _, e := syscall.Syscall(procGetConsoleMode.Addr(), 2, uintptr(fd), uintptr(unsafe.Pointer(&st)), 0)
+	r, _, e := syscall.Syscall(procGetConsoleMode.Addr(), 2, fd, uintptr(unsafe.Pointer(&st)), 0)
 	return r != 0 && e == 0
 }
